+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
-**Last Updated:** 2026-03-12
+**Last Updated:** 2026-03-26
+
+---
+
+## [0.4.2 - 2026-03-26]
+
+### Updates
+- Added scratches to pre-game threads
+
+### Bugfixes
+- Fixed a bug where post-game threads would be prevented from posting if game day threads weren't also enabled
+- Potential fix to lost threads/threads never posting (modified existing scheduled thread check)
+
+### Misc.
+- Added r/OnlyHabs to allowlist
+
+### Devvit
+- Update to devvit 0.12.15
 
 ---
 
@@ -20,6 +37,9 @@
 - Fixed an issue where extra attacker ("EA") goals would be displayed with the power play ("PP") strength modifier
 - Fixed an issue where live updates could stop before the game was over
 - Unsticky should now also remove from new reddit highlights
+
+### Devvit
+- Update to devvit 0.12.14
 
 ---